Attempt to fix intermittent initial sizing issue
My best guess for this is a race condition between rendering reflow (100dvh) and the observed height. If the JS code wins, it will take an incorrect measurement which isn't corrected until we get a resize event which may never occur. Instead this tries using a ResizeObserver to directly measure what we care about (the shader-container changing size)